requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
netmiko
python-nmap
requests